DF-44
* Refactor commonalities into ChunkMessage
