+++ src/main/java/messaging/ChunkReadRequest.java
@@ -4,11 +4,7 @@ import java.io.DataInputStream;
 import java.io.DataOutputStream;
 import java.io.IOException;
 
-public class ChunkReadRequest extends Message {
-
-    public String absoluteFilePath;
-
-    public Integer sequence;
+public class ChunkReadRequest extends ChunkMessage {
 
     public ChunkReadRequest(String hostname, String ipAddress, Integer port, String absoluteFilePath, Integer sequence) {
         this.hostname = hostname;
@@ -30,38 +26,6 @@ public class ChunkReadRequest extends Message {
     @Override
     public MessageType getType() {
         return MessageType.CHUNK_READ_REQUEST;
-    }
-
-    public String getAbsoluteFilePath() {
-        return absoluteFilePath;
-    }
-
-    public Integer getSequence() {
-        return sequence;
-    }
-
-    /**
-     * In addition to the header, writes the chunk's filename and sequence.
-     * @param dataOutputStream The DataOutputStream we are writing to.
-     * @throws IOException If fails to read to DataOutputStream
-     */
-    @Override
-    public void marshal(DataOutputStream dataOutputStream) throws IOException {
-        super.marshal(dataOutputStream); // first marshal common Message header
-        writeString(dataOutputStream, this.absoluteFilePath);
-        dataOutputStream.writeInt(this.sequence);
-    }
-
-    /**
-     * In addition to the header, reads the chunk's filename and sequence.
-     * @param dataInputStream The DataInputStream we are reading from.
-     * @throws IOException If fails to read from DataInputStream
-     */
-    @Override
-    public void unmarshal(DataInputStream dataInputStream) throws IOException {
-        super.unmarshal(dataInputStream); // first unmarshal common Message header
-        this.absoluteFilePath = readString(dataInputStream);
-        this.sequence = dataInputStream.readInt();
     }
 
     @Override
