+++ src/main/java/messaging/ChunkReplicateCommand.java
@@ -9,16 +9,10 @@ import java.io.IOException;
  * to a target Chunk Server for replication. Happens when the replication of a chunk
  * falls below a certain level due to Chunk Server failure.
  */
-public class ChunkReplicateCommand extends Message {
+public class ChunkReplicateCommand extends ChunkMessage {
 
     // The hostname of the Chunk Server the chunk replica should be stored at
     public String targetChunkServer;
-
-    // The absolute path of the chunk that needs to be replicated
-    public String absoluteFilePath;
-
-    // The sequence index of the chunk that needs to be replicated
-    public Integer sequence;
 
     public ChunkReplicateCommand(String hostname, String ipAddress, Integer port, String targetChunkServer,
                                  String absoluteFilePath, Integer sequence) {
@@ -48,17 +42,9 @@ public class ChunkReplicateCommand extends Message {
         return targetChunkServer;
     }
 
-    public String getAbsoluteFilePath() {
-        return absoluteFilePath;
-    }
-
-    public Integer getSequence() {
-        return sequence;
-    }
-
     /**
-     * In addition to the header, writes the target Chunk Server's hostname where the replica should be stored,
-     * as well as the chunk filename and sequence that should be replicated
+     * In addition to the header, filename, and sequence of the chunk, writes the target Chunk Server's hostname where
+     * the replica should be stored.
      * @param dataOutputStream The DataOutputStream we are writing to.
      * @throws IOException If fails to write to DataOutputStream
      */
@@ -66,21 +52,19 @@ public class ChunkReplicateCommand extends Message {
     public void marshal(DataOutputStream dataOutputStream) throws IOException {
         super.marshal(dataOutputStream); // first marshal common Message header
         writeString(dataOutputStream, this.targetChunkServer);
-        writeString(dataOutputStream, this.absoluteFilePath);
-        dataOutputStream.writeInt(this.sequence);
+
     }
 
     /**
-     * In addition to the header, reads the Chunk Server's hostname where another replica is stored.
+     * In addition to the header, filename, and sequence of the chunk, reads the Chunk Server's hostname where another
+     * replica should be stored.
      * @param dataInputStream The DataInputStream we are reading from.
      * @throws IOException If fails to read from DataInputStream
      */
     @Override
     public void unmarshal(DataInputStream dataInputStream) throws IOException {
-        super.unmarshal(dataInputStream); // first unmarshal common Message header
+        super.unmarshal(dataInputStream); // first unmarshal common Message header=
         this.targetChunkServer = readString(dataInputStream);
-        this.absoluteFilePath = readString(dataInputStream);
-        this.sequence = dataInputStream.readInt();
     }
 
     @Override
@@ -89,17 +73,17 @@ public class ChunkReplicateCommand extends Message {
         if (other == this) return true;
         if (!(other instanceof ChunkReplicateCommand)) return false;
         ChunkReplicateCommand crcOther = (ChunkReplicateCommand) other;
-        return (this.targetChunkServer.equals(crcOther.getTargetChunkServer()) &&
-                this.absoluteFilePath.equals(crcOther.getAbsoluteFilePath()) &&
-                this.sequence.equals(crcOther.getSequence()));
+        return (this.absoluteFilePath.equals(crcOther.getAbsoluteFilePath()) &&
+                this.sequence.equals(crcOther.getSequence()) &&
+                this.targetChunkServer.equals(crcOther.getTargetChunkServer()));
     }
 
     @Override
     public String toString() {
         return "ChunkReplicateCommand:" +
-                String.format("\n  targetChunkServer: %s", this.targetChunkServer) +
                 String.format("\n  absoluteFilePath: %s", this.absoluteFilePath) +
-                String.format("\n  sequence: %d", this.sequence);
+                String.format("\n  sequence: %d", this.sequence) +
+                String.format("\n  targetChunkServer: %s", this.targetChunkServer);
     }
