+++ src/main/java/messaging/ClientWriteRequest.java
@@ -8,13 +8,7 @@ import java.io.IOException;
  * A Message concrete class which should be targeted at the Controller to request a list of Chunk Servers
  * to write a chunk to.
  */
-public class ClientWriteRequest extends Message {
-
-    // The name of the file we are attempting to write
-    public String absoluteFilePath;
-
-    // The sequence index of the chunk within the file we are writing
-    public Integer sequence;
+public class ClientWriteRequest extends ChunkMessage {
 
     public ClientWriteRequest(String hostname, String ipAddress, Integer port, String absoluteFilePath, Integer sequence) {
         this.hostname = hostname;
@@ -38,39 +32,6 @@ public class ClientWriteRequest extends Message {
         return MessageType.CLIENT_WRITE_REQUEST;
     }
 
-    public String getAbsoluteFilePath() {
-        return absoluteFilePath;
-    }
-
-    public Integer getSequence() {
-        return sequence;
-    }
-
-    /**
-     * In addition to the header, writes the absolute path of the file for the chunk we
-     * are requesting to write, and the sequence number of the chunk within that file.
-     * @param dataOutputStream The DataOutputStream we are writing to.
-     * @throws IOException If fails to read to DataOutputStream
-     */
-    @Override
-    public void marshal(DataOutputStream dataOutputStream) throws IOException {
-        super.marshal(dataOutputStream); // first marshal common Message header
-        writeString(dataOutputStream, this.absoluteFilePath);
-        dataOutputStream.writeInt(this.sequence);
-    }
-
-    /**
-     * In addition to the header, reads the absolute path of the file for the chunk being
-     * requested for writing, and the sequence number of the chunk within that file.
-     * @param dataInputStream The DataInputStream we are reading from.
-     * @throws IOException If fails to read from DataInputStream
-     */
-    @Override
-    public void unmarshal(DataInputStream dataInputStream) throws IOException {
-        super.unmarshal(dataInputStream); // first unmarshal common Message header
-        this.absoluteFilePath = readString(dataInputStream);
-        this.sequence = dataInputStream.readInt();
-    }
 
     @Override
     public boolean equals(Object other) {
